+++ crates/chain-listener/src/event/mod.rs
@@ -17,11 +17,12 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-pub mod cc_activated;
-pub mod compute_unit_matched;
+mod cc_activated;
+mod compute_unit_matched;
 mod unit_activated;
 mod unit_deactivated;
 
-pub use compute_unit_matched::ComputeUnitMatched;
+pub use cc_activated::CommitmentActivated;
+pub use compute_unit_matched::{ComputeUnitMatched, CIDV1};
 pub use unit_activated::UnitActivated;
 pub use unit_deactivated::UnitDeactivated;
