+++ crates/chain-listener/src/lib.rs
@@ -24,11 +24,11 @@
 #![feature(result_option_inspect)]
 extern crate core;
 
-pub use event::cc_activated::CommitmentActivated;
-pub use event::compute_unit_matched::CIDV1;
+pub use event::CommitmentActivated;
 pub use event::ComputeUnitMatched;
 pub use event::UnitActivated;
 pub use event::UnitDeactivated;
+pub use event::CIDV1;
 pub use listener::ChainListener;
 
 mod event;
@@ -36,3 +36,4 @@ mod listener;
 
 mod persistence;
 mod proof_tracker;
+mod types;
